PLS function performance assessment using \(MSEP\) indicator

The perf.PLS function allows to assess PLS models using\(MSEP\) criterion. This function returns a list including \(MSEP\) values for each number of components. perf.PLS gives also a suggestion about the number of components selection. A plot allows to visualize the model performance according to the number of components.

Usage

perf.PLS(object, ncomp = object$ncomp, K=nrow(object$X))

Arguments

object

Object of class inheriting from "pls".

ncomp

The number of components desired for MSEP computing (no more than the number of components computed in the model).

K

The number of blocks for cross-validation (leave-one-out by default).

Details

\(K\) must be a value between 2 and the the number of rows of the dataset used for training model.

MSEP is defined by mean of squared error between the true \(Y\) values and their associated predictions.

See also

Examples


library(sgPLSdevelop)

## data and model creation 
d <- data.create(p = 10, list = TRUE)
n <- nrow(d$X)
ncomp.max <- 10
X <- d$X
Y <- d$Y
modele <- PLS(X,Y,ncomp = ncomp.max, mode = "regression")

## using perf.PLS function
perf.res <- perf.PLS(modele)
#> Error in perf.PLS(modele): object 'ncomp.max' not found
perf.res$MSEP 
#> Error: object 'perf.res' not found
perf.res$h.best #number of components suggestion 
#> Error: object 'perf.res' not found