Performance PLS function assessment using Q2 indicator.

The q2.PLS function allows to assess PLS models using \(Q2\) criterion. This function returns a list including \(Q2\) values for each number of components. A plot allows to visualize the model performance according to the number of components.

Usage

q2.PLS(object, mode = "regression", ncomp.max = object$ncomp)

Arguments

object

Object of class inheriting from "pls".

comp.max

The number of components desired for q2 computing (no more than the number of components computed in the model).

mode

Character string. What type of algorithm to use, (partially) matching one of "regression" or "canonical". See Details.

Details

A suggestion about the number of components selection is also given by Q2 criterion. Endly, other indicators are given such as PRESS and RSS for each number of components and such as \(PRESSj\) and \(RSSj\) (matrices) for a given number of components and for a given column of Y dataset.

Value

\(PRESSj\) and \(RSSj\) are matrices of size \(ncomp.max\) x \(q\). The row sums of \(PRESSj\) and \(RSSj\) give respectively PRESS and RSS.

References

K.-A. Le Cao, Zoe Welham, Multivariate data integration using R (pages 172 to 174), MixOmics

See also

Examples


library(sgPLSdevelop)
#> Error in library(sgPLSdevelop): there is no package called 'sgPLSdevelop'

d <- data.create(n=50, q=2)
#> Error in data.create(n = 50, q = 2): could not find function "data.create"
X <- d$X
#> Error: object 'd' not found
Y <- d$Y
#> Error: object 'd' not found

ncomp.max <- 10
model.pls <- PLS(X = X, Y = Y, ncomp = ncomp.max, mode = "regression")
#> Error in PLS(X = X, Y = Y, ncomp = ncomp.max, mode = "regression"): could not find function "PLS"

q2.res <- q2.PLS(model.pls, ncomp.max = ncomp.max, mode = "regression")
#> Error in q2.PLS(model.pls, ncomp.max = ncomp.max, mode = "regression"): could not find function "q2.PLS"

# q2 values
q2.res$q2
#> Error: object 'q2.res' not found

# PRESS values
q2.res$PRESS
#> Error: object 'q2.res' not found

# RSS values
q2.res$RSS
#> Error: object 'q2.res' not found

# PRESS values by column
q2.res$PRESSj
#> Error: object 'q2.res' not found

# RSS values by column
q2.res$RSSj
#> Error: object 'q2.res' not found